backend/requirements.txt
# FastAPI + ASGI
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Database
sqlalchemy>=2.0.0
alembic>=1.14.0
psycopg2-binary>=2.9.9
pgvector>=0.3.0

# Async tasks
celery[redis]>=5.4.0
redis>=5.0.0

# PDF processing
pymupdf>=1.25.0
pytesseract>=0.3.13
Pillow>=11.0.0

# NLP / Embedding
httpx>=0.28.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Validation
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
httpx>=0.28.0

# Linting
ruff>=0.8.0
